ストア周りの処理の一部をstore.tsに移した
